incident location gets saved

+++ src/main/java/com/example/Terriffic/Incident/Repository/IncidentDao.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface IncidentDao extends JpaRepository<Incident, Integer> {
+public interface IncidentDao extends JpaRepository<Incident, Long> {
     @Query(value = "SELECT * FROM Incident i WHERE ST_DWithin(i.location, ST_MakePoint(:lonMin, :latMin), ST_Distance(ST_MakePoint(:lonMin, :latMin), ST_MakePoint(:lonMax, :latMax)))", nativeQuery = true)
     List<Incident> findWithinBoundingBox(
             @Param("latMin") double latMin,
