+++ src/main/java/com/example/Terriffic/Incident/Service/IncidentService.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface IncidentService {
     public Incident save(Incident incident);
-    public void deleteById(int id);
+    public void deleteById(Long id);
     public List<Incident> findAll();
-    public Incident findById(int id);
+    public Incident findById(Long id);
     public List<Incident> findWithinBoundingBox(double latMin, double lonMin, double latMax, double lonMax);
 }
